refactor(PAT-65): Move files. Estruct directory.
