+++ com/frontend/app/src/containers/Card.jsx
@@ -10,7 +10,7 @@ import { Draggable } from 'react-beautiful-dnd';
 import Card from '@mui/material/Card';
 import CardContent from '@mui/material/CardContent';
 import Typography from '@mui/material/Typography';
-import ConsutationDisplay from './ConsutationDisplay.jsx'
+import ConsutationDisplay from '../components/ConsutationDisplay.jsx'
 
 
 const CustomCard = ({ card, index }) => {
