+++ com/frontend/app/src/pages/CaseTakerPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import Dashboard from '../containers/Dashboard';
-import Consultancy from '../components/Consultancy';
+import Consultancy from '../containers/Consultancy';
 
 const theme = createTheme();
 export default function CaseTaker( ) {
