+++ com/frontend/app/src/utils/panel.jsx
